annabel/requirements.txt
telethon
supabase==2.15.3
python-dotenv